+++ etna/transforms/math/scalers.py
@@ -16,6 +16,11 @@ class StandardScalerTransform(SklearnTransform):
     """Standardize features by removing the mean and scaling to unit variance.
 
     Uses sklearn.preprocessing.StandardScaler inside.
+
+    Warning
+    -------
+    This transform can suffer from look-ahead bias. For transforming data at some timestamp
+    it uses information from the whole train part.
     """
 
     def __init__(
@@ -64,9 +69,14 @@ class StandardScalerTransform(SklearnTransform):
 
 
 class RobustScalerTransform(SklearnTransform):
-    """
-    Scale features using statistics that are robust to outliers.
+    """Scale features using statistics that are robust to outliers.
+
     Uses sklearn.preprocessing.RobustScaler inside.
+
+    Warning
+    -------
+    This transform can suffer from look-ahead bias. For transforming data at some timestamp
+    it uses information from the whole train part.
     """
 
     def __init__(
@@ -132,9 +142,14 @@ class RobustScalerTransform(SklearnTransform):
 
 
 class MinMaxScalerTransform(SklearnTransform):
-    """
-    Transform features by scaling each feature to a given range.
+    """Transform features by scaling each feature to a given range.
+
     Uses sklearn.preprocessing.MinMaxScaler inside.
+
+    Warning
+    -------
+    This transform can suffer from look-ahead bias. For transforming data at some timestamp
+    it uses information from the whole train part.
     """
 
     def __init__(
@@ -183,9 +198,14 @@ class MinMaxScalerTransform(SklearnTransform):
 
 
 class MaxAbsScalerTransform(SklearnTransform):
-    """
-    Scale each feature by its maximum absolute value.
+    """Scale each feature by its maximum absolute value.
+
     Uses sklearn.preprocessing.MaxAbsScaler inside.
+
+    Warning
+    -------
+    This transform can suffer from look-ahead bias. For transforming data at some timestamp
+    it uses information from the whole train part.
     """
 
     def __init__(