+++ etna/transforms/missing_values/imputation.py
@@ -161,6 +161,11 @@ class TimeSeriesImputerTransform(PerSegmentWrapper):
     - This transform can't fill NaNs in the future, only on train data.
     - This transform can't fill NaNs in non-zero strategy if all values are Nans. In this case exception is raised.
     - In 'forward_fill' strategy very first value and first NaNs are replaced with zero.
+
+    Warning
+    -------
+    This transform can suffer from look-ahead bias in 'mean' mode. For transforming data at some timestamp
+    it uses information from the whole train part.
     """
 
     def __init__(self, in_column: str = "target", strategy: str = ImputerMode.zero, window: int = -1):