+++ etna/transforms/missing_values/resample.py
@@ -98,7 +98,13 @@ class _OneSegmentResampleWithDistributionTransform(Transform):
 
 
 class ResampleWithDistributionTransform(PerSegmentWrapper):
-    """ResampleWithDistributionTransform resamples the given column using the distribution of the other column."""
+    """ResampleWithDistributionTransform resamples the given column using the distribution of the other column.
+
+    Warning
+    -------
+    This transform can suffer from look-ahead bias. For transforming data at some timestamp
+    it uses information from the whole train part.
+    """
 
     def __init__(
         self, in_column: str, distribution_column: str, inplace: bool = True, out_column: Optional[str] = None